requirements.txt
openai-agents==0.1.8
slack_sdk==3.27.0
gspread==6.0.2
pydantic==2.7.1
Flask==3.0.3
python-dotenv==1.0.1 